MAX_DATA_SIZE was causing segmentation fault
There is a macro called MAX_DATA_SIZE defined as dependend from the number of points in the simulation.
But, sometimes we can have one more point, so it was trying to acess a memory that was not even
allocated, causing the segmentation fault. It explains also why I was having problems when I was trying to
free this region of memory.